+++ Diary/Extension/UITableViewCell+.swift
@@ -1,0 +1,12 @@
+//
+//  UITableViewCell+.swift
+//  Diary
+//
+//  Created by dudu, papri on 2022/06/14.
+//
+
+import UIKit
+
+extension UITableViewCell {
+    static let reuseIdentifier = String(describing: UITableViewCell.self)
+}
